rust/ffi: move AppLayerEvent to ffi
Ticket: 7666
