+++ rust/derive/src/applayerevent.rs
@@ -52,12 +52,12 @@ pub fn derive_app_layer_event(input: TokenStream) -> TokenStream {
     // Suricata name space as "suricata". Check the CARGO_PKG_NAME environment variable to
     // determine what identifier to setup.
     let is_suricata = std::env::var("CARGO_PKG_NAME")
-        .map(|var| var == "suricata")
+        .map(|var| var == "suricata_ffi")
         .unwrap_or(false);
     let crate_id = if is_suricata {
         syn::Ident::new("crate", proc_macro2::Span::call_site())
     } else {
-        syn::Ident::new("suricata", proc_macro2::Span::call_site())
+        syn::Ident::new("suricata_ffi", proc_macro2::Span::call_site())
     };
 
     let expanded = quote! {
@@ -91,7 +91,7 @@ pub fn derive_app_layer_event(input: TokenStream) -> TokenStream {
             unsafe extern "C" fn get_event_info(
                 event_name: *const std::os::raw::c_char,
                 event_id: *mut u8,
-                event_type: *mut #crate_id::core::AppLayerEventType,
+                event_type: *mut #crate_id::applayer::AppLayerEventType,
             ) -> std::os::raw::c_int {
                 #crate_id::applayer::get_event_info::<#name>(event_name, event_id, event_type)
             }
@@ -99,7 +99,7 @@ pub fn derive_app_layer_event(input: TokenStream) -> TokenStream {
             unsafe extern "C" fn get_event_info_by_id(
                 event_id: u8,
                 event_name: *mut *const std::os::raw::c_char,
-                event_type: *mut #crate_id::core::AppLayerEventType,
+                event_type: *mut #crate_id::applayer::AppLayerEventType,
             ) -> std::os::raw::c_int {
                 #crate_id::applayer::get_event_info_by_id::<#name>(event_id, event_name, event_type)
             }
