+++ rust/ffi/src/applayer.rs
@@ -18,6 +18,8 @@
 //! App-layer utils.
 
 use crate::cast_pointer;
+use std::ffi::CStr;
+pub use suricata_sys::sys::AppLayerEventType;
 use suricata_sys::sys::{
     AppLayerGetTxIterState, AppLayerGetTxIterTuple, AppLayerResult, AppProto, StreamSlice,
 };
@@ -226,4 +228,107 @@ pub unsafe extern "C" fn state_get_tx_iterator<S: State<Tx>, Tx: Transaction>(
 ) -> AppLayerGetTxIterTuple {
     let state = cast_pointer!(state, S);
     state.get_transaction_iterator(min_tx_id, &mut (*istate).un.u64_)
+}
+
+// Make the AppLayerEvent derive macro available to users importing
+// AppLayerEvent from this module.
+pub use suricata_derive::AppLayerEvent;
+
+/// AppLayerEvent trait that will be implemented on enums that
+/// derive AppLayerEvent.
+pub trait AppLayerEvent {
+    /// Return the enum variant of the given ID.
+    fn from_id(id: u8) -> Option<Self>
+    where
+        Self: std::marker::Sized;
+
+    /// Convert the enum variant to a C-style string (suffixed with \0).
+    fn to_cstring(&self) -> &str;
+
+    /// Return the enum variant for the given name.
+    fn from_string(s: &str) -> Option<Self>
+    where
+        Self: std::marker::Sized;
+
+    /// Return the ID value of the enum variant.
+    fn as_u8(&self) -> u8;
+
+    /// # Safety
+    ///
+    /// Caller should provide a valid cstring for event_name.
+    /// a non-null pointer for event_id
+    /// and a non-null pointer for event_type
+    unsafe extern "C" fn get_event_info(
+        event_name: *const std::os::raw::c_char, event_id: *mut u8,
+        event_type: *mut AppLayerEventType,
+    ) -> std::os::raw::c_int;
+
+    /// # Safety
+    ///
+    /// Caller should provide
+    /// a non-null pointer for event_name
+    /// and a non-null pointer for event_type
+    unsafe extern "C" fn get_event_info_by_id(
+        event_id: u8, event_name: *mut *const std::os::raw::c_char,
+        event_type: *mut AppLayerEventType,
+    ) -> std::os::raw::c_int;
+}
+
+/// Generic `get_info_info` implementation for enums implementing
+/// AppLayerEvent.
+///
+/// Normally usage of this function will be generated by
+/// derive(AppLayerEvent), for example:
+///
+/// ```rust,ignore
+/// #[derive(AppLayerEvent)]
+/// enum AppEvent {
+///     EventOne,
+///     EventTwo,
+/// }
+///
+/// get_event_info::<AppEvent>(...)
+/// ```
+///
+/// # Safety
+///
+/// Caller should provide a valid cstring for event_name.
+/// a non-null pointer for event_id
+/// and a non-null pointer for event_type
+#[inline(always)]
+pub unsafe fn get_event_info<T: AppLayerEvent>(
+    event_name: *const std::os::raw::c_char, event_id: *mut u8, event_type: *mut AppLayerEventType,
+) -> std::os::raw::c_int {
+    if event_name.is_null() {
+        return -1;
+    }
+
+    let event = match CStr::from_ptr(event_name).to_str().map(T::from_string) {
+        Ok(Some(event)) => event.as_u8(),
+        _ => {
+            return -1;
+        }
+    };
+    *event_type = AppLayerEventType::APP_LAYER_EVENT_TYPE_TRANSACTION;
+    *event_id = event;
+    0
+}
+
+/// Generic `get_info_info_by_id` implementation for enums implementing
+/// AppLayerEvent.
+/// # Safety
+///
+/// Caller should provide
+/// a non-null pointer for event_name
+/// and a non-null pointer for event_type
+#[inline(always)]
+pub unsafe fn get_event_info_by_id<T: AppLayerEvent>(
+    event_id: u8, event_name: *mut *const std::os::raw::c_char, event_type: *mut AppLayerEventType,
+) -> std::os::raw::c_int {
+    if let Some(e) = T::from_id(event_id) {
+        *event_name = e.to_cstring().as_ptr() as *const std::os::raw::c_char;
+        *event_type = AppLayerEventType::APP_LAYER_EVENT_TYPE_TRANSACTION;
+        return 0;
+    }
+    -1
 }
